Fix Storage bug
Timetable related JsonAdapted classes did
not have annotations for the name of the variables
in the constructor. This causes errors when trying to load
existing JSON files into TimetaBRO.

Fixed errors by adding @JsonProperty labels to the parameters
in the within the constructors in the following files:
* JsonAdaptedCca
* JsonAdaptedDatedEvent
* JsonAdaptedModule
* JsonAdaptedMeetUpEvent

Removed JsonAdaptedTimeBlock due to redundancy.