Bug fix: Exclusive end time for FreeTime
Schedule sees end time for TimeBlocks as exclusive. If an event is from 1200 - 1230, schedule recognises that the free time before that ends at 1130, not 1200.

Let's:
- make endTime of freetime to be inclusive
- extend the FreeTime and TimeBlock implementation to include accepting 2400 as the endTime.

Now the implementation of FreeTime and TimeBlock allows a range of time intervals starting from 0000 and ending at 2400.

This allows events to end at the latest possible timing, which is actually 12am the next day.