Fix Edit and EditUser
Edit and EditUser does not check for duplicate phone numbers
and emails. Add, Edit, and EditUser also did not check if numbers
and emails were the same as the user.

Fix Edit and EditUser commands with guard clauses checking for
duplicate phones.
Fix hasPhone, hasEmail, and hasPerson to include user in its checks.
Add tests checking for duplicate emails and numbers to both edit commands.
Add overloaded methods in EditPersonDescriptorBuilder and
EditUserDescriptorBuilder